# [WIP] resurrecting in place `ntoh`