Resurrection of https://github.com/tamasgal/UnROOT.jl/pull/101 (which means some copy-pasting from @Moelf ;) ). Basically, we add a `inplace::Bool` to `interped_data` which toggles between doing `ntoh.(reinterpret(...))` (copy) or an inplace-variant, both returning the same type which means it's still type stable.

But with a small modification, `basketarray()`'s return type depends on the boolean. Probably this can be fixed with a barrier function to discard  the `Ref{UInt8[]}` if called with `inplace=false` (the default.

Unit tests all pass.

```julia
julia> using UnROOT

julia> const t = LazyTree(ROOTFile("uncompressed_Run2012BC_DoubleMuParked_Muons.root"),"Events");

julia> function bar(t)
           for evt in t
               _ = evt.Muon_pt
           end
           nothing
       end
```

### before
```julia
julia> @btime sum(t.nMuon) seconds=10
  204.417 ms (6078 allocations: 469.98 MiB)
0x0000000008e67ad8

julia> @btime bar(t) seconds=10
  712.140 ms (14304 allocations: 1.82 GiB)
```

### after

```julia
julia> @btime sum(t.nMuon) seconds=10
  205.316 ms (6780 allocations: 235.24 MiB)
0x0000000008e67ad8

julia> @btime bar(t) seconds=10
  725.626 ms (16241 allocations: 1.26 GiB)
```

Performance is pretty much the same but with less total allocated memory. If I profile `bar(t)` I see there's a materialization. I thought this should be in place??

![image](https://user-images.githubusercontent.com/5760027/133875328-2fef9e25-0083-4603-8047-79ca750b9dbc.png)