@Moelf 2021-09-18
actually:
```julia
#before
(14304 allocations: 1.82 GiB)
#after
(16241 allocations: 1.26 GiB)
```

I think this means we shouldn't do this, the number of allocation increases probably caused slow down. And the difference isn't as big as the 2x in the `nMuon` case