@Moelf 2021-12-02
I wonder if @oschulz has any opinion. In fact, it would be nice to have a variation of `VectorOfVector` that has a field for `Ref` to hold the original byte array...