@oschulz 2021-12-02
I'm not sure I've understood all the implications here - are you looking for somethings like a `VectorOfVectors` wrapped around an [`UnsafeArray`](https://github.com/JuliaArrays/UnsafeArrays.jl)?