@Moelf 2021-12-02
the problem is we have an original byte array `A::Vector{UInt8}`, we want to wrap `VoV` around it except we want to "cast" it to, for example, `B::Vector{Int32}`, but if you cast it by `unsafe_wrap()`, GC thinks `A` is not needed and thus GC-ed and you get bad data in `VoV(B)`.