@oschulz 2021-12-02
But why can't we use `ReinterpretArray` if we have to keep the reference to `A` around anyway?