@Moelf 2021-12-02
because `ReinterpretArray` is slow for both:
1. `A .= ntoh.(A)`
2. everything user will later on do, such as looping over the branch