@Moelf 2022-02-12
https://github.com/JuliaLang/julia/issues/42227#issuecomment-1037463738

on Julia master, `y .= ntoh.(y)` is no longer slow given `y` is an `ReinterpretedArray`