@aminnj 2022-02-13
It might be that zlib decompression + the calculation is washing out the true benchmark